third_party/blink/web_tests/pending_beacon/pending_beacon-basic.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<script>
  test(() => {
    assert_throws_js(TypeError, () => new PendingBeacon('/'));
  }, `PendingBeacon cannot be constructed directly.`);

  const BeaconTypes = [
    {type: PendingPostBeacon, name: "PendingPostBeacon", expectedMethod: "POST"},
    {type: PendingGetBeacon, name: "PendingGetBeacon", expectedMethod: "GET"},
  ];

  for (const beaconType of BeaconTypes) {
    test(() => {
      assert_throws_js(TypeError, () => new beaconType.type());
    }, `${beaconType.name}: constructor throws TypeError if no url is provided`);

    test(() => {
      const beacon = new beaconType.type("/");
      assert_equals(beacon.url, "/");
      assert_equals(beacon.method, beaconType.expectedMethod);
      assert_equals(beacon.backgroundTimeout, -1);
      assert_equals(beacon.timeout, -1);
      assert_true(beacon.pending);
    }, `${beaconType.name}: constructor sets default properties when not provided.`);

    test(() => {
      const beacon = new beaconType.type("https://www.google.com",
                      {backgroundTimeout: 200, timeout: 100});
      assert_equals(beacon.url, "https://www.google.com");
      assert_equals(beacon.method, beaconType.expectedMethod);
      assert_equals(beacon.backgroundTimeout, 200);
      assert_equals(beacon.timeout, 100);
      assert_true(beacon.pending);
    }, `${beaconType.name}: constructor can set properties.`);

    test(() => {
      let beacon = new beaconType.type("https://www.google.com",
                      {method: "GET", backgroundTimeout: 200, timeout: 100});

      beacon.backgroundTimeout = 400;
      assert_equals(beacon.backgroundTimeout, 400);

      beacon.timeout = 600;
      assert_equals(beacon.timeout, 600);

    }, `${beaconType.name}: backgroundTimeout & timeout can be mutated.`);

    test(() => {
      let beacon = new beaconType.type("https://www.google.com");

      beacon.url = "/";
      assert_equals(beacon.url, "https://www.google.com");

      beacon.method = "FOO";
      assert_equals(beacon.method, beaconType.expectedMethod);

      assert_true(beacon.pending);
      beacon.pending = false;
      assert_true(beacon.pending);
    }, `${beaconType.name}: url & method & pending cannot be mutated.`);
  }

</script>
</body>
</html>
